refactor: use nokogiri not selenium
Previously we used selenium to extract both plain text and HTML
elements, because selenium was better than nokogiri for plain text, and
it could do elements too. We had to switch to pandoc to extract plain
text, so we can now switch to nokogiri for elements, which will be
easier to maintain and probably much faster. We can also implement it as
a Cloud Function rather than Cloud Run, because it doesn't require any
system dependencies.
