+++ .github/workflows/html-to-text-lint-and-test.yml
@@ -1,14 +1,14 @@
-name: Ruby Lint and Test
+name: Html-to-text lint and test
 
 on:
   workflow_dispatch:
   pull_request:
     paths:
-      - 'docker/parse-html/**'
+      - 'docker/html-to-text/**'
 
 defaults:
   run:
-    working-directory: docker/parse-html
+    working-directory: docker/html-to-text
 
 jobs:
   build:
