+++ .github/workflows/parse-html-lint-and-test.yml
@@ -1,0 +1,37 @@
+name: Parse-html lint and test
+
+on:
+  workflow_dispatch:
+  pull_request:
+    paths:
+      - 'src/cloud-functions/parse-html/**'
+
+defaults:
+  run:
+    working-directory: src/cloud-functions/parse-html
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v3
+
+    # Set up Ruby
+    - name: Set up Ruby
+      uses: ruby/setup-ruby@v1
+      with:
+        ruby-version: '3.2.2' # or the version you need
+
+    # Install dependencies
+    - name: Install dependencies
+      run: |
+        bundle install --jobs 4 --retry 3
+
+    # Run RuboCop
+    - name: Run RuboCop
+      run: bundle exec rubocop
+
+    # Run RSpec
+    - name: Run RSpec tests
+      run: bundle exec rspec
