+++ src/cloud-functions/parse-html/Gemfile
@@ -2,7 +2,7 @@ source "https://rubygems.org"
 
 gem "functions_framework"
 gem "json"
-gem "selenium-webdriver"
+gem "nokogiri"
 
 group :development, :test do
   gem "byebug", require: true
