+++ terraform-dev/parse-html.tf
@@ -1,32 +1,62 @@
-# A service to use as a remote function in BigQuery
-# Then create a place to put the app images
-resource "google_cloud_run_v2_service" "parse_html" {
-  name     = "parse-html"
-  location = var.region
-  ingress  = "INGRESS_TRAFFIC_INTERNAL_ONLY"
+# A BigQuery remote function
+data "archive_file" "parse_html" {
+  type        = "zip"
+  output_path = "/tmp/parse-html.zip"
+  source_dir  = "../src/cloud-functions/parse-html"
+}
 
-  template {
-    containers {
-      image = "europe-west2-docker.pkg.dev/${var.project_id}/${google_artifact_registry_repository.docker.repository_id}/parse-html:latest"
-      resources {
-        limits = {
-          cpu    = "1000m"  # If we put "1" or nothing, terraform reapplies it.
-          memory = "2048Mi" # By experiment, necessary and sufficient.
-        }
+resource "google_storage_bucket_object" "parse_html" {
+  name   = "sourcecode.zip"
+  bucket = google_storage_bucket.cloud_functions.name
+  source = data.archive_file.parse_html.output_path # Add path to the zipped function source code
+}
+
+resource "google_cloudfunctions2_function" "parse_html" {
+  name        = "parse-html"
+  location    = var.region
+  description = "Extract HTML elements from GOV.UK content"
+
+  build_config {
+    runtime     = "ruby32"
+    entry_point = "parse_html" # Set the entry point
+    source {
+      storage_source {
+        bucket = google_storage_bucket.cloud_functions.name
+        object = google_storage_bucket_object.parse_html.name
       }
     }
-    # The function only handles one request at a time, which it enforces by
-    # using a filesystem lock as a mutex. It might be worth telling GCP
-    # explicitly, so that it might try to create more instances of the function.
-    max_instance_request_concurrency = 1
+    docker_repository = "projects/${var.project_id}/locations/${var.region}/repositories/gcf-artifacts"
+  }
+
+  service_config {
+    available_memory = "512Mi"
+    # available_cpu = 1 # This function is CPU-bound and serial, so we need exactly one
+    timeout_seconds    = 60
+    max_instance_count = 100
+    # max_instance_request_concurrency = 1
+  }
+
+  # Force terraform to redeploy the function when the source code changes
+  # https://github.com/hashicorp/terraform-provider-google/issues/1938#issuecomment-1229042663
+  lifecycle {
+    replace_triggered_by = [
+      google_storage_bucket_object.parse_html
+    ]
   }
 }
 
-resource "google_bigquery_connection" "parse_html" {
-  connection_id = "parse-html"
-  description   = "Remote function parse_html"
-  location      = var.region
-  cloud_resource {}
+data "google_iam_policy" "cloud_function_parse_html" {
+  binding {
+    role = "roles/cloudfunctions.invoker"
+    members = [
+      "serviceAccount:${google_bigquery_connection.parse_html.cloud_resource[0].service_account_id}",
+    ]
+  }
+}
+
+resource "google_cloudfunctions2_function_iam_policy" "parse_html" {
+  cloud_function = google_cloudfunctions2_function.parse_html.name
+  policy_data    = data.google_iam_policy.cloud_function_parse_html.policy_data
 }
 
 data "google_iam_policy" "cloud_run_parse_html" {
@@ -40,14 +70,40 @@ data "google_iam_policy" "cloud_run_parse_html" {
 
 resource "google_cloud_run_v2_service_iam_policy" "parse_html" {
   location    = var.region
-  name        = google_cloud_run_v2_service.parse_html.name
+  name        = google_cloudfunctions2_function.parse_html.name
   policy_data = data.google_iam_policy.cloud_run_parse_html.policy_data
+}
+
+resource "google_bigquery_connection" "parse_html" {
+  connection_id = "parse-html"
+  description   = "Remote function parse_html"
+  location      = var.region
+  cloud_resource {}
+}
+
+data "google_iam_policy" "bigquery_connection_parse_html" {
+  binding {
+    role = "roles/bigquery.connectionUser"
+    members = [
+      google_service_account.bigquery_scheduled_queries.member,
+    ]
+  }
+}
+
+resource "google_bigquery_connection_iam_policy" "parse_html" {
+  connection_id = google_bigquery_connection.parse_html.connection_id
+  policy_data   = data.google_iam_policy.bigquery_connection_parse_html.policy_data
 }
 
 # generate a random string suffix for a bigquery job to deploy the function
 resource "random_string" "deploy_parse_html" {
   length  = 20
   special = false
+  lifecycle {
+    replace_triggered_by = [
+      google_storage_bucket_object.parse_html
+    ]
+  }
 }
 
 ## Run a bigquery job to deploy the remote function
@@ -62,7 +118,7 @@ resource "google_bigquery_job" "deploy_parse_html" {
       {
         project_id = var.project_id
         region     = var.region
-        uri        = google_cloud_run_v2_service.parse_html.uri
+        uri        = google_cloudfunctions2_function.parse_html.url
       }
     )
     create_disposition = "" # must be set to "" for scripts
@@ -71,7 +127,7 @@ resource "google_bigquery_job" "deploy_parse_html" {
 
   lifecycle {
     replace_triggered_by = [
-      google_cloud_run_v2_service.parse_html
+      google_storage_bucket_object.parse_html
     ]
   }
 }
